Update patch set 3
Patch Set 3:

I think we should evaluate if the flow is like that. The if statement should execute just on cloud and else just on fog? I think so. But I could not see a flow that it could.

Patch-set: 3
Reviewer: Ramon Henrique <1013096@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0