Don't consider 404 an error in get devices
When a gateway doesn't have a device, it returns an error with code 404.
As it was being treated as a request error, a get devices request would
always return an empty list of devices if an empty gateway is connect to
the cloud.

This patch also avoids getGatewayDevices to call its callback multiple
times.

Change-Id: I93bc3a37000e6cf6d5a16c548af05317eb977e79
Reviewed-on: https://review.gerrithub.io/419747
Reviewed-by: João Neto <joao.alexandre@cesar.org.br>
Reviewed-by: Ramon Henrique <rhpr@cin.ufpe.br>
Reviewed-by: Thiago Cardoso <thiago.figueredo@cesar.org.br>
Tested-by: Thiago Cardoso <thiago.figueredo@cesar.org.br>