Don't consider 404 an error in get devices
When a gateway doesn't have a device, it returns an error with code 404.
As it was being treated as a request error, a get devices request would
always return an empty list of devices if an empty gateway is connect to
the cloud.

This patch also avoids getGatewayDevices to call its callback multiple
times.